Move `char_buffer` into a separate file
